Create Original script, Garlic bugged
So this script allows you to define certain objects as display. As you put a certain amount of items in a container, the same amount of display locations will be toggled on (if you have made enough). The reverse is true on item removal. There is also a function that forbids you from placing any items in the linked container, without adding them in a Formlist. To fully set up this script, it can take a while, but should be easily done.